pipeline/requirements.txt
anthropic==0.84.0
supabase==2.28.0
feedparser==6.0.12
requests==2.32.5
python-dotenv==1.2.1
beautifulsoup4==4.12.3